% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{fitted_adastrumm}
\alias{fitted_adastrumm}
\title{Fitted values from flexible models for longitudinal data}
\usage{
fitted_adastrumm(mod)
}
\arguments{
\item{mod}{The fitted model, output from fit_adastrumm}
}
\description{
Fitted values from flexible models for longitudinal data
}
