% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_adastrumm}
\alias{predict_adastrumm}
\title{Prediction for AdaStruMMs}
\usage{
predict_adastrumm(
  mod,
  newdata,
  deriv = FALSE,
  interval = FALSE,
  level = 0.95,
  samples = NULL,
  n_samples = 1000
)
}
\arguments{
\item{mod}{The fitted model, output from fit_adastrumm}

\item{newdata}{A data frame, containing the new
values of x, y and c at which to make predictions.}

\item{deriv}{If TRUE, output predictions for derivative
of subject-specific mean function}

\item{interval}{Should a confidence interval also be given
for the predictions?}

\item{level}{The coverage for the confidence interval,
if interval = TRUE.}

\item{samples}{An optional list of parametric bootstrap samples
used to construct the confidence interval,
output by find_samples.}

\item{n_samples}{If samples = NULL, the number of parametric
bootstrap samples used to construct the confidence interval,
if interval = TRUE.}
}
\value{
If interval = FALSE, a vector of the predictions.
If interval = TRUE, a data frame with columns estimate,
lower and upper, giving the point estimate and
confidence interval
}
\description{
Prediction for AdaStruMMs
}
\examples{
data_full <- simulate_1dv(1, -0.5, 0.1, 0.5, 0.1, 20, 10)
data <- data_full$data
mod <- fit_adastrumm(data)
pred_data <- data_full$pred_data
pred_data$mu_c_hat <- predict_adastrumm(mod, newdata = pred_data, interval = TRUE)
}
