% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_adastrumm}
\alias{update_adastrumm}
\title{Update an Adaptively-Structure Mixed Model (AdaStruMM)}
\usage{
update_adastrumm(data, prev_fit, fix_tuning = TRUE, fix_pop_par = FALSE)
}
\arguments{
\item{data}{A data frame, with columns c (identifying the
individual subjects), x (the time) and y (the response).}

\item{prev_fit}{a previous model fit, fit from a subset of the data.}

\item{fix_tuning}{logical: should the tuning parameters K and gamma be fixed?}

\item{fix_pop_par}{logical: should the population-level parameters be fixed?}
}
\value{
The fitted model.
}
\description{
Update an Adaptively-Structure Mixed Model (AdaStruMM)
}
