% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_adastrumm}
\alias{fit_adastrumm}
\title{Fit an Adaptively-Structure Mixed Model (AdaStruMM)}
\usage{
fit_adastrumm(
  data,
  nbasis = 10,
  kmax = 10,
  k_tol = 1e-04,
  lsp_poss = -5:15,
  trace = FALSE,
  alpha_index = 1,
  normalise = TRUE
)
}
\arguments{
\item{data}{A data frame, with columns c (identifying the
individual subjects), x (the time) and y (the response).}

\item{nbasis}{The number of spline basis functions.}

\item{kmax}{The maximum number of functional principal components
to allow.}

\item{lsp_poss}{The grid of possible values to consider for
log(gamma), the log of the smoothing parameter.}

\item{trace}{If TRUE, print out extra information.}

\item{k_tol.}{The tolerance to use in selecting k, to explain at
least 1-k_tol of the variation in trajectories.}
}
\value{
The fitted model.
}
\description{
Fit an Adaptively-Structure Mixed Model (AdaStruMM)
}
\examples{
data_full <- simulate_1dv(1, -0.5, 0.1, 0.5, 0.1, 20, 10)
data <- data_full$data
mod <- fit_adastrumm(data)
}
